feature(chart): Сделал отрисовку нажатия на часть диаграммы

+++ lib/units/operationes_stats/cubit/change_stats_state.dart
@@ -18,7 +18,7 @@ class ChangeStatsError extends ChangeStatsState {
 class ChangeStatsLoaded extends ChangeStatsState {
   final StatsDateType dateType;
   final Map<int, List<ItemOperationModel>> operationes;
-  final int summOfElement;
+  final double summOfElement;
 
   ChangeStatsLoaded(
     this.dateType,
